
\name{brca.tcga}
\docType{data}
\alias{brca.tcga}
\title{A sample breast cancer gene differential expression analysis t statistic score data}
\description{
  This data set gives the TCGA breast cancer tumor versus normal t-test scores of each gene.
}
\usage{brca.tcga
}
\format{A matrix containing the t statistics of 17960 genes}
\source{TCGA}
\keyword{datasets}
