
\name{drug.info}
\docType{data}
\alias{drug.info}
\title{The matrix containing the drug instances information of CMap drug data}
\description{
  This matrix gives the drug instance information of CMap drug data, including drug names, instance numbers, cell line, dosage, etc.
}
\usage{drug.info
}
\format{A matrix of 6100 rows and 7 columns.}
\source{CMap}
\keyword{datasets}
