
\name{example.disease}
\docType{data}
\alias{example.disease}
\title{An example gene differential expression analysis test statistic score data}
\description{
  This data gives an example of gene differential expression analysis test statistic score as the inpur for DrInsight analysis.
}
\usage{example.disease
}
\format{A matrix containing the simulated scores of 1000 genes}
\source{simulation}
\keyword{datasets}
