
\name{example.drug.profiles}
\docType{data}
\alias{example.drug.profiles}
\title{An example cmap drug reference profiles}
\description{
  This data gives an simulated example of cmap drug perturbed reference profiles.
}
\usage{example.drug.profiles
}
\format{An object containing the simulated reference drug rank matrix and drug information table}
\source{simulation}
\keyword{datasets}
