
\name{probe.to.genes}
\docType{data}
\alias{probe.to.genes}
\title{The matrix of Affymetrix probe IDs and gene symbols relationship}
\description{
  This matrix gives the resource to convert the Affymetrix probe IDs in CMap data into gene symbols.
}
\usage{probe.to.genes
}
\format{A matrix of two columns and 12994 rows}
\source{Affymetrix}
\keyword{datasets}
